Move logic from controller of user and img
Moved the logic of the ProfilesController to UserService, ProductByBasketService and ProductService, also, moved the logic from AbstractProductFieldsController to AbstractImagesService

+++ src/main/java/com/example/GeekShop/controller/product_fields/AbstractProductFieldsController.java
@@ -6,14 +6,12 @@ import com.example.GeekShop.model.product_fields.AbstractProductField;
 import com.example.GeekShop.repository.images_product_fields.ImageRepository;
 import com.example.GeekShop.repository.product_fields.AbstractProductFieldRepository;
 import com.example.GeekShop.service.images_product_fields.AbstractImagesService;
-import com.example.GeekShop.service.product.ProductService;
 import com.example.GeekShop.service.product_fields.AbstractProductFieldService;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -147,43 +145,45 @@ public abstract class AbstractProductFieldsController<
         getService().deleteById(id);
         return "redirect:/" + url;
     }
-
     @Override
     public String formEditElement(@PathVariable Long id, @NonNull Model model) {
         model.addAttribute("element", getService().findById(id));
         model.addAttribute("url", url);
         return "product_fields/edit_element";
     }
-
-    private TI toImageEntity(MultipartFile file) throws IOException {
-        TI image = getImage();
-        image.setName(file.getName());
-        image.setOriginalName(file.getOriginalFilename());
-        image.setContentType(file.getContentType());
-        image.setSize(file.getSize());
-        image.setBytes(file.getBytes());
-        return image;
-    }
     protected List<Product> productsByProductFields(E element) {
         return null;
     }
+    private TI convertFromMultipartFileToImage(MultipartFile file) throws IOException {
+        TI image = getImage();
+        getImageService().setFieldsForImage(image, file);
+        return image;
+    }
     private void saveImages(MultipartFile file1, MultipartFile file2, MultipartFile file3, E element) {
-        element.setPreviewsId(null);
-        element.setImages(new ArrayList<>());
-        getImageService().deleteAllByElement(element);
-        try {
-            TI image1 = toImageEntity(file1);
-            image1.setIsPreviews(true);
-            element.addImages(image1, element);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+        if (file1 != null && file2 != null && file3 != null && element != null) {
+            element.setPreviewsId(null);
+            element.setImages(new ArrayList<>());
+            getImageService().deleteAllByElement(element);
+            try {
+                TI image1 = convertFromMultipartFileToImage(file1);
+                image1.setIsPreviews(true);
+                element.addImages(image1, element);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+            try {
+                element.addImages(convertFromMultipartFileToImage(file2), element);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+            try {
+                element.addImages(convertFromMultipartFileToImage(file3), element);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+            getService().save(element);
+            element.setPreviewsId(element.getImages().get(0).getId());
+            getService().save(element);
         }
-        try { element.addImages(toImageEntity(file2), element); }
-        catch (IOException e) { throw new RuntimeException(e); }
-        try { element.addImages(toImageEntity(file3), element); }
-        catch (IOException e) { throw new RuntimeException(e); }
-        getService().save(element);
-        element.setPreviewsId(element.getImages().get(0).getId());
-        getService().save(element);
     }
 }
